run the task before registering the task definition
This is because we use GITHUB_TOKEN as a task environment
variable. Environment variables are stored in the clear in
a task definition. When the task definition is registered, the token
is available in the task definition.

This is okay because the token expires at the end of a GitHub Actions
job. However, if the task definition is registered first (and the token is exposed) and
the task takes a while to run then a valid token is exposed for the
duration of the long-running task.

This commit runs the task first. Immediately after the task finishes
running, the task definition is registered. Since this action is the
last step of the job, the token expires after the registration completes.
The token exposd in the task definition is invalid.